+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## v1.1.5 (1.20.1)
-- fixed banner material translations
-- fixed other bugs relating to armory update
+- added new expiremental Alt-render
+- Altrenderer should be fully compatible with Iris, Oculus and Optifine